% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdols.R
\docType{data}
\name{marksul2003}
\alias{marksul2003}
\title{Sample panel data for pdols}
\format{A data frame with 760 rows and 5 variables:}
\source{
\url{http://www.utdallas.edu/~d.sul/papers/}
}
\usage{
marksul2003
}
\description{
Sample panel data in long format for pdols
}
\keyword{datasets}

