% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdols.R
\docType{package}
\name{pdolsms}
\alias{pdolsms}
\alias{pdolsms-package}
\title{pdolsms: Cointegration Vector Estimation by Panel DOLS}
\description{
The pdolsms package provides a function for the estimation of panel data cointegrating relationships following Mark and Sul (2003), <DOI:10.1111/j.1468-0084.2003.00066.x>.
}

