% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdols.R
\name{summary.pdols}
\alias{summary.pdols}
\title{A summary method for the pdols class}
\usage{
\method{summary}{pdols}(object, ...)
}
\arguments{
\item{object}{An object of class pdols}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
A summary method for the pdols class
}
\examples{
data("marksul2003")
result <- pdols(formula = Y ~ X + Z - 1,
                index = c("Country", "Year"),
                data = marksul2003,
                p = 2,
                icase = "constant")
summary(result)
}
\references{
Mark, N. C. and Sul, D. (2003), Cointegration Vector Estimation by Panel DOLS and Long-run Money Demand. Oxford Bulletin of Economics and Statistics, 65: 655-680. <DOI:10.1111/j.1468-0084.2003.00066.x>
}

